+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Tooltip to filters
 
+## [3.33.1] - 2019-10-04
+### Fixed
+-  problems with the `fetchMore` that is used on the useFetchMore by handling possible errors
+
 ## [3.33.0] - 2019-10-02
 ### Added
 - Loading more results of a search result now changes the url. The back and forward feature is working properly to consider the page the user was last in.